ansible-galaxy - do not require mandatory keys when listing collections
If the galaxy.yml file does not contain mandatory keys, display a warning and '*'
for the version number.

This fix feels very leaky. List may be better off using it's own method to get metadata
about a role rather than relying on existing functions. Many of the existing
functions enforce strict standards, which make sense for everything but list,
which should be robust and just display information.